#1 Initial commit of models, scraper, pre-processing scripts, and data
This commit contains the initial commit of many things.

We have our webscraper. It's implemented using pyppeteer. It's slow, but it works. IT gives us the ability to click buttons on pages if we want to go that route.

We have our initial model in this commit as well. It's a simple LSTM.

This patch also contains the data pre-processing utilities we'll use.