#3 Get website up and running
This commit more or less gets the website for the text generator up and running.

We have a space background. That's pretty cool.

We have buttons linking to our website and GitHub profile.

We have a text box to display text.

We have a nice button to generate new text.

The functionality is still dummy functionality.

We need to get it to show real data.